requirements.txt
snowflake-connector-python
pandas
pyarrow<19.0.0
toml